Group related CSS values
