+++ packages/alfa-css/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./syntax";
 export * from "./value";
-export * from "./value/angle";
 export * from "./value/box";
 export * from "./value/calculation";
 export * from "./value/color";
@@ -10,18 +9,13 @@ export * from "./value/color/hsl";
 export * from "./value/color/named";
 export * from "./value/color/rgb";
 export * from "./value/color/system";
-export * from "./value/converter";
-export * from "./value/dimension";
+export * from "./value/unit/converter";
 export * from "./value/gradient";
 export * from "./value/gradient/linear";
 export * from "./value/gradient/radial";
 export * from "./value/image";
-export * from "./value/integer";
 export * from "./value/keyword";
-export * from "./value/length";
-export * from "./value/number";
 export * from "./value/numeric";
-export * from "./value/percentage";
 export * from "./value/position";
 export * from "./value/shadow";
 export * from "./value/shape";
@@ -39,5 +33,5 @@ export * from "./value/transform/rotate";
 export * from "./value/transform/scale";
 export * from "./value/transform/skew";
 export * from "./value/transform/translate";
-export * from "./value/unit";
+export * from "./value/unit/unit";
 export * from "./value/url";
