+++ packages/alfa-css/src/value/numeric/number.ts
@@ -1,7 +1,7 @@
 import { Parser } from "@siteimprove/alfa-parser";
 import { Slice } from "@siteimprove/alfa-slice";
 
-import { Token } from "../syntax/token";
+import { Token } from "../../syntax/token";
 import { Numeric } from "./numeric";
 
 const { map } = Parser;
@@ -59,9 +59,8 @@ export namespace Number {
   /**
    * {@link https://drafts.csswg.org/css-values/#number-value}
    */
-  export const parse: Parser<
-    Slice<Token>,
-    Number,
-    string
-  > = map(Token.parseNumber(), (number) => Number.of(number.value));
+  export const parse: Parser<Slice<Token>, Number, string> = map(
+    Token.parseNumber(),
+    (number) => Number.of(number.value)
+  );
 }
