+++ packages/alfa-css/src/value/numeric/percentage.ts
@@ -1,7 +1,7 @@
 import { Parser } from "@siteimprove/alfa-parser";
 import { Slice } from "@siteimprove/alfa-slice";
 
-import { Token } from "../syntax/token";
+import { Token } from "../../syntax/token";
 import { Numeric } from "./numeric";
 
 const { map } = Parser;
@@ -52,11 +52,8 @@ export namespace Percentage {
     return value instanceof Percentage;
   }
 
-  export const parse: Parser<
-    Slice<Token>,
-    Percentage,
-    string
-  > = map(Token.parsePercentage(), (percentage) =>
-    Percentage.of(percentage.value)
+  export const parse: Parser<Slice<Token>, Percentage, string> = map(
+    Token.parsePercentage(),
+    (percentage) => Percentage.of(percentage.value)
   );
 }
