+++ packages/alfa-css/src/value/transform/skew.ts
@@ -5,9 +5,8 @@ import { Slice } from "@siteimprove/alfa-slice";
 import { Token } from "../../syntax/token";
 import { Value } from "../../value";
 
-import { Angle } from "../angle";
-import { Number } from "../number";
-import { Unit } from "../unit";
+import { Angle, Number } from "../numeric";
+import { Unit } from "../unit/unit";
 
 const { map, left, right, pair, either, delimited, option } = Parser;
