+++ packages/alfa-css/tsconfig.json
@@ -12,7 +12,6 @@
     "src/syntax/nth.ts",
     "src/syntax/token.ts",
     "src/value.ts",
-    "src/value/angle.ts",
     "src/value/box.ts",
     "src/value/calculation.ts",
     "src/value/color.ts",
@@ -22,18 +21,19 @@
     "src/value/color/named.ts",
     "src/value/color/rgb.ts",
     "src/value/color/system.ts",
-    "src/value/converter.ts",
-    "src/value/dimension.ts",
     "src/value/gradient.ts",
     "src/value/gradient/linear.ts",
     "src/value/gradient/radial.ts",
     "src/value/image.ts",
-    "src/value/integer.ts",
     "src/value/keyword.ts",
-    "src/value/length.ts",
-    "src/value/number.ts",
     "src/value/numeric.ts",
-    "src/value/percentage.ts",
+    "src/value/numeric/angle.ts",
+    "src/value/numeric/dimension.ts",
+    "src/value/numeric/integer.ts",
+    "src/value/numeric/length.ts",
+    "src/value/numeric/number.ts",
+    "src/value/numeric/numeric.ts",
+    "src/value/numeric/percentage.ts",
     "src/value/position.ts",
     "src/value/shadow.ts",
     "src/value/shape.ts",
@@ -52,6 +52,8 @@
     "src/value/transform/skew.ts",
     "src/value/transform/translate.ts",
     "src/value/unit.ts",
+    "src/value/unit/converter.ts",
+    "src/value/unit/unit.ts",
     "src/value/url.ts",
     "test/syntax/declaration.spec.ts",
     "test/syntax/lexer.spec.ts",
